Refactoring of voice play/record components
